git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13808 01821946-851b-0410-b7f4-f1477c70e57a

+++ 2d/string.cpp
@@ -991,7 +991,7 @@ int32_t StringWidth (char * s, int32_t n)
 	char		p;
 
 if (n > 0) {
-	if (n >= strlen (s)) 
+	if (n >= (int32_t) strlen (s)) 
 		n = 0;
 	else {
 		p = s [n];
